removed unused functions in benchmark tests
