+++ app/src/androidTest/java/com/example/app/ExampleInstrumentedTest.kt
@@ -2,7 +2,6 @@ package com.example.app
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import dagger.hilt.android.testing.HiltAndroidTest
 
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -14,7 +13,6 @@ import org.junit.Assert.*
  *
  * See [testing documentation](http://d.android.com/tools/testing).
  */
-@HiltAndroidTest
 @RunWith(AndroidJUnit4::class)
 class ExampleInstrumentedTest {
     @Test
